pyproject.toml
[project]
name = "competitor_watcher"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "mag1cfrog", email = "harrywong2017@gmail.com"},
]
dependencies = [
    "python-dotenv",
    "polars",
    "duckdb",
    "pyarrow",
    "loguru>=0.7.2",
    "python-amazon-sp-api @ git+https://github.com/mag1cfrog/python-amazon-sp-api.git@master",
    "pytz>=2024.2",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
